// src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
      categories: {
        Row: {
          database_id: number | null;
          id: number;
          name: string;
        };
        Insert: {
          database_id?: number | null;
          id?: number;
          name: string;
        };
        Update: {
          database_id?: number | null;
          id?: number;
          name?: string;
        };
        Relationships: [
          {
            foreignKeyName: "public_categories_database_id_fkey";
            columns: ["database_id"];
            isOneToOne: false;
            referencedRelation: "databases";
            referencedColumns: ["id"];
          }
        ];
      };
      databases: {
        Row: {
          documentation: string;
          id: number;
          name: string;
          schema: string;
        };
        Insert: {
          documentation: string;
          id?: number;
          name: string;
          schema: string;
        };
        Update: {
          documentation?: string;
          id?: number;
          name?: string;
          schema?: string;
        };
        Relationships: [];
      };
      tasks: {
        Row: {
          category_id: number | null;
          database_id: number | null;
          id: number;
          row_strict: boolean;
          solution: string;
          text: string;
        };
        Insert: {
          category_id?: number | null;
          database_id?: number | null;
          id?: number;
          row_strict?: boolean;
          solution: string;
          text: string;
        };
        Update: {
          category_id?: number | null;
          database_id?: number | null;
          id?: number;
          row_strict?: boolean;
          solution?: string;
          text?: string;
        };
        Relationships: [
          {
            foreignKeyName: "public_tasks_database_id_fkey";
            columns: ["database_id"];
            isOneToOne: false;
            referencedRelation: "databases";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "tasks_category_id_fkey";
            columns: ["category_id"];
            isOneToOne: false;
            referencedRelation: "categories";
            referencedColumns: ["id"];
          }
        ];
      };
      user_progress: {
        Row: {
          completed: boolean | null;
          id: number;
          task_id: number | null;
          user_id: string | null;
        };
        Insert: {
          completed?: boolean | null;
          id?: number;
          task_id?: number | null;
          user_id?: string | null;
        };
        Update: {
          completed?: boolean | null;
          id?: number;
          task_id?: number | null;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "public_user_progress_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "user_progress_task_id_fkey";
            columns: ["task_id"];
            isOneToOne: false;
            referencedRelation: "tasks";
            referencedColumns: ["id"];
          }
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      requesting_user_id: {
        Args: Record<PropertyKey, never>;
        Returns: string;
      };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
// ...
};

type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
      PublicSchema["Views"])
  ? (PublicSchema["Tables"] &
      PublicSchema["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R;
    }
    ? R
    : never
  : never;